database/postgresql/requirements.txt
psycopg2-binary
click
requests
